==> modeling/freight_generation_model/code/Step_00_Read_NAICS_Codes.py <==
import pandas as pd
import numpy as np


df = pd.read_csv('input/CF1700A08.dat', sep='|')

df = df.loc[df['YEAR']==2017]

df_ = df.loc[df['GEOTYPE']==1]

df_grp = df_.groupby(['NAICS2012', 'NAICS2012_TTL'], as_index=False)['TON'].sum()


# =============================================================================
# naics in CFS
# =============================================================================
# NAICS_in_CFS = ['212', '311', '312', '313', '314', '315', '316',
#        '321', '322', '323', '324', '325', '326', '327', '331', '332',
#        '333', '334', '335', '336', '337', '339', '4231',
#        '4232', '4233', '4234', '4235', '4236', '4237', '4238', '4239',
#        '4241', '4242', '4243', '4244', '4245', '4246', '4247',
#        '4248', '4249', '4541', '45431', '4931', '5111', '551114']

df_grp.to_csv('output/step_0/naics_codes.csv', index=False)
